Address PR comments: move utility tests to test_utils.py and update CHANGELOG
