+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- Refactored `BellmanPeriod` with type hints, docstrings, and improved parameter
+  handling
+- Introduced `_resolve_parameters`, `_resolve_decision_rules`, and
+  `_resolve_shocks` helper methods for consistent fallback logic
+- Added gradient computation methods (`grad_reward_function`,
+  `grad_transition_function`, `grad_pre_state_function`) to `BellmanPeriod`
+- Added shock resolution support in `BellmanPeriod` methods
+- Moved `compute_gradients_for_tensors` tests from `test_bellman.py` to
+  `test_utils.py`
+
 ### Added
 
 - Initial release features
